refactor: CQS 패턴 적용
