+++ src/main/java/com/nhnacademy/store99/bookstore/cart/service/CartQueryService.java
@@ -1,0 +1,8 @@
+package com.nhnacademy.store99.bookstore.cart.service;
+
+import com.nhnacademy.store99.bookstore.cart.dto.response.CartItemResponse;
+import java.util.List;
+
+public interface CartQueryService {
+    List<CartItemResponse> getCartItemsByUser();
+}
